apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.10.0,<3.0
pydantic-settings>=2.10.0,<3.0
email-validator>=2.0.0
PyJWT==2.8.0
aiosqlite==0.20.0
python-multipart==0.0.12
pytest>=8.3.0
pytest-asyncio>=0.21.0
httpx>=0.27.0
